fix(cli): address P1 review + CI failures — defer repoRoot, remove global.d.ts, add registry entry
Fixes three issues found in CI and Codex review:

1. P1 Codex review — repoRoot computed at module import time (scripts/lib/repo.ts:46):
   - Extract findRepoRoot into scripts/lib/find-repo-root.ts with no module-level side effects.
   - scripts/cli.ts now imports from lib/find-repo-root.js, bypassing the eager repoRoot
     evaluation that triggered before --cwd could be parsed. The --cwd flag now works
     correctly when invoked from a directory with no package.json or .git.
   - scripts/lib/repo.ts imports + re-exports findRepoRoot; its repoRoot const is still
     computed eagerly (as expected for scripts run in-repo), just no longer triggered by cli.ts.

2. TS2688 CI failure — global.d.ts re-introduced the same broken resolution as "types":["node"]:
   - Remove scripts/global.d.ts. The /// <reference types="node" /> directive uses the same
     type-library loader path as "types":["node"] in tsconfig, both of which fail with
     TypeScript 6.0.3 + @types/node@25.x. TypeScript's auto-discovery via typeRoots is
     the correct mechanism and requires no explicit directive.
   - Also remove "types":["node"] from tsconfig.typedoc.json for the same reason.
   - Revert the typedoc.json exclude entry for the now-deleted global.d.ts.

3. check:automation-registry failure — check:cli-contract npm script not in registry:
   - Add check:cli-contract entry to tools/automation-registry.yml after check:agents.

Also updates check-cli-contract.ts to verify find-repo-root.ts exists and exports
findRepoRoot (was checking for export function in repo.ts, now adapted for re-export).
Adds docs/scripts/lib/find-repo-root/ stubs and modules.md entry for the new module.
