+++ docs/scripts/lib/find-repo-root/README.md
@@ -1,0 +1,17 @@
+---
+title: "lib/find-repo-root"
+folder: "docs/scripts/lib/find-repo-root"
+description: "Entry point for generated API reference for the lib/find-repo-root script helper module."
+entry_point: true
+---
+[**@luis85/agentic-workflow**](../../README.md)
+
+***
+
+[@luis85/agentic-workflow](../../modules.md) / lib/find-repo-root
+
+# lib/find-repo-root
+
+## Functions
+
+- [findRepoRoot](functions/findRepoRoot.md)
