+++ tsconfig.typedoc.json
@@ -6,8 +6,7 @@
     "target": "ES2022",
     "strict": true,
     "noEmit": true,
-    "skipLibCheck": true,
-    "types": ["node"]
+    "skipLibCheck": true
   },
   "include": ["scripts/**/*.ts"]
 }
